requirement.txt
django==3.2.0
asgiref==3.3.4
markupsafe==1.1.1
pytz==2021.1
sqlparse==0.4.1
requests==2.25.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
mysqlclient==2.0.3
redis==3.5.3
uwsgi==2.0.19.1
django-filter==2.4.0
uvicorn==0.13.4
werkzeug==1.0.1
jwt==1.2.0
jinja2==2.11.3